alteracoes para adequar a feature de pedidos ao desenho de arquitetura

+++ app/src/main/java/br/com/appfastfood/produto/aplicacao/adaptadores/requisicao/ProdutoRequisicao.java
@@ -1,6 +1,5 @@
 package br.com.appfastfood.produto.aplicacao.adaptadores.requisicao;
 
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -10,7 +9,7 @@ import java.math.BigDecimal;
 public class ProdutoRequisicao {
     private Long id;
     private String nome;
-    private BigDecimal preco;
+    private Double preco;
     private String uriImagem;
     private String categoria;
     private String descricao;
